Cast the value of sizeof as unsigned long for the purposes of portable output formatting. (dm)
git-svn-id: svn://mielke.cc/main/brltty@2954 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/charset.c
@@ -215,8 +215,8 @@ getWcharCharset (void) {
 
   if (!wcharCharset) {
     char charset[0X10];
-    snprintf(charset, sizeof(charset), "UCS-%u%cE",
-             sizeof(wchar_t),
+    snprintf(charset, sizeof(charset), "UCS-%lu%cE",
+             (unsigned long)sizeof(wchar_t),
 #ifdef WORDS_BIGENDIAN
              'B'
 #else /* WORDS_BIGENDIAN */
